extracting filterBounties function
